Add initial BenchFlowTest definition; extracts properties that are common between Test and Experiment in extensible protocol; adjusts tests accordingly
