+++ src/test/scala/cloud/benchflow/test/config/experiment/PropertiesSpec.scala
@@ -1,16 +1,17 @@
 package cloud.benchflow.test.config.experiment
 
+import cloud.benchflow.test.config.ConfigurationYamlProtocol
 import org.scalatest.{Matchers, FlatSpec}
 
 /**
   * @author Simone D'Avico (simonedavico@gmail.com)
   *
-  *         Created on 18/07/16.
+  * Created on 18/07/16.
   */
-class PropertiesSpec extends FlatSpec with Matchers {
+class PropertiesSpec extends FlatSpec with Matchers with ConfigurationYamlProtocol {
 
   import net.jcazevedo.moultingyaml._
-  import BenchFlowExperimentYamlProtocol._
+  import cloud.benchflow.test.config._
 
   "Properties" should "parse correctly" in {
